% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Behavioral_profile.R
\name{activity_multiple_times_executed}
\alias{activity_multiple_times_executed}
\title{activity sometimes multiple times executed}
\usage{
activity_multiple_times_executed(repetition_and_path_log, xml_internal_doc,
  activity, direct_parallel)
}
\arguments{
\item{repetition_and_path_log}{repetition and path log list object created by the function create_repetition_and_path_log}

\item{xml_internal_doc}{document object created using the create_internal_document function}

\item{activity}{the activity name}

\item{direct_parallel}{a table containing the direct and parallel relations}
}
\value{
a boolean value indicating whether it is true that an activity can be executed multiple times in the same path
}
\description{
{ This functions returns true or false on whether or not an activity is sometimes multiple times executed
This can be useful for measuring the understandability using behavioral profiles.
}
}
\examples{
\dontrun{activity_multiple_times_executed(log, doc, "A")}
}
