% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Behavioral_profile.R
\name{activity_names_repetitions}
\alias{activity_names_repetitions}
\title{activity names repetitions}
\usage{
activity_names_repetitions(repetition_and_path_log, xml_internal_doc)
}
\arguments{
\item{repetition_and_path_log}{repetition and path log list object created by the function create_repetition_and_path_log}

\item{xml_internal_doc}{document object created using the create_internal_document function}
}
\value{
a list containing the repetitions with their respective activity names
}
\description{
{ This functions returns a list containing the repetitions with their respective activity names
This can be useful for measuring the understandability using behavioral profiles.
}
}
\examples{
\dontrun{activity_multiple_times_executed(log, doc, "A")}
}
