% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{avg_connector_degree}
\alias{avg_connector_degree}
\title{Average connector degree}
\usage{
avg_connector_degree(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the average connector degree
}
\description{
Average connector degree is defined as the average incoming and outgoing sequence flows of all gateways and activities with at least two incoming or outgoing sequence flows
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
avg_connector_degree(file_path)
}
