% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{coefficient_network_connectivity}
\alias{coefficient_network_connectivity}
\title{Coefficient of network connectivity}
\usage{
coefficient_network_connectivity(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the coefficient of network connectivity
}
\description{
Coefficient of network connectivity is defined as the number of sequence flows divided by the size
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
coefficient_network_connectivity(file_path)
}
