% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cognitive_weight}
\alias{cognitive_weight}
\title{Cognitive weights}
\usage{
cognitive_weight(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the control flow complexity
}
\description{
{Cognitive weight is defined as a weighted sum of gateways and activities}
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
cognitive_weight(file_path)
}
