% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{connectivity_level_between_pools}
\alias{connectivity_level_between_pools}
\title{The connectivity level between pools}
\usage{
connectivity_level_between_pools(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the connectivity level between pools
}
\description{
The connectivity level between pools is the number of message flows over the number of pools
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
connectivity_level_between_pools(file_path)
}
