% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{control_flow_complexity}
\alias{control_flow_complexity}
\title{Control flow complexity}
\usage{
control_flow_complexity(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the control flow complexity
}
\description{
{Control flow complexity is defined as the sum of the outgoing of exclusive gateways, the number of parallel gateways
and two to the power of all outgoing sequence flows of the inclusive gateways}
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
control_flow_complexity(file_path)
}
