% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{coupling_metric}
\alias{coupling_metric}
\title{Coupling metric}
\usage{
coupling_metric(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the control flow complexity
}
\description{
{Coupling metric is defined as the sum of the number of activities, AND-splits and a weighterd number of OR and XOR splits}
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
coupling_metric(file_path)
}
