% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_internal_doc}
\alias{create_internal_doc}
\title{A function for creating internal documents}
\usage{
create_internal_doc(bpmn_file, signavio = FALSE)
}
\arguments{
\item{bpmn_file}{file path of the BPMN file}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an object containing the xml document
}
\description{
Is used for creating xml documents which nearly every function of this package needs as an input
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
create_internal_doc(file_path)
}
