% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cyclicity}
\alias{cyclicity}
\title{Cyclicity}
\usage{
cyclicity(repetition_and_path_log, xml_internal_doc, signavio = FALSE)
}
\arguments{
\item{repetition_and_path_log}{repetition and path log list object created by the function create_repetition_and_path_log}

\item{xml_internal_doc}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the cyclicity
}
\description{
Cyclicity is defined as the number of nodes on a cycle divided by the total number of nodes
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")
path_log <- create_path_and_repetition_log(create_internal_doc(file_path))}
cyclicity(path_log, create_internal_doc(file_path))
}
