% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{density_process_model}
\alias{density_process_model}
\title{Density}
\usage{
density_process_model(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the density
}
\description{
Density is defined as the number of sequence flows divided by the size times the size minus one
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
density_process_model(file_path)
}
