% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{depth}
\alias{depth}
\title{Depth}
\usage{
depth(repetition_and_path_log)
}
\arguments{
\item{repetition_and_path_log}{repetition and path log list object created by the function create_repetition_and_path_log}
}
\value{
an integer indicating the depth
}
\description{
Depth is defined as the the nesting of the process model.
If there is a split gateway, the depth is increased with one.
If there is a join gateway, the depth is decreased with one.
The cumulative sum is taken and the maximum of the cumulative sum is calculated for each path.
The nesting depth is the maximum of each path value
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")
path_log <- create_path_and_repetition_log(create_internal_doc(file_path))}
depth(path_log)
}
