% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{diameter}
\alias{diameter}
\title{Diameter}
\usage{
diameter(repetition_and_path_log)
}
\arguments{
\item{repetition_and_path_log}{repetition and path log list object created by the function create_repetition_and_path_log}
}
\value{
an integer indicating the diameter
}
\description{
Length of longest path, in practice the length of longest path.
The assumption is made that one repetition for each loop is allowed and these repetitions count as well for the diameter
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")
path_log <- create_path_and_repetition_log(create_internal_doc(file_path))}
diameter(path_log)
}
