% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Behavioral_profile.R
\name{direct_parallel_relations}
\alias{direct_parallel_relations}
\title{Direct and parallel relations}
\usage{
direct_parallel_relations(repetition_and_path_log, xml_internal_doc)
}
\arguments{
\item{repetition_and_path_log}{repetition and path log list object created by the function create_repetition_and_path_log}

\item{xml_internal_doc}{document object created using the create_internal_document function}
}
\value{
a table as described in the description
}
\description{
{ This functions returns a table containing all direct and parallel relations between activities. The table contains five columns:
- the two first represent activity ids
- the third represents the type of relations, which is parallel or direct
- the last two columns are the corresponding activity names for the first two columns
}
}
\examples{
\dontrun{direct_parallel_relations(repetition_and_path_log, xml_internal_doc)}
}
