% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Behavioral_profile.R
\name{filtered_path_log_parallel}
\alias{filtered_path_log_parallel}
\title{Filter path log with only traces containing the parallel gateway together with the relevant activity}
\usage{
filtered_path_log_parallel(structured_path_log, xml_internal_doc,
  activity_name)
}
\arguments{
\item{structured_path_log}{repetition and path log list object created by the function create_repetition_and_path_log}

\item{xml_internal_doc}{document object created using the create_internal_document function}

\item{activity_name}{name of the activity for the relevant filtering}
}
\value{
the filtered path log
}
\description{
{ This functions returns a path log with no traces with a parallel gateway of which the given activity is part but not included
}
}
\examples{
\dontrun{direct_parallel_relations(repetition_and_path_log, xml_internal_doc)}
}
