% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{max_connector_degree}
\alias{max_connector_degree}
\title{Maximum connector degree}
\usage{
max_connector_degree(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the maximum connector degree
}
\description{
Maximum connector degree is defined as the gateway or activity with the most incoming and outgoing sequence flows
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
max_connector_degree(file_path)
}
