% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{n_data_objects}
\alias{n_data_objects}
\title{Data Objects}
\usage{
n_data_objects(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the number of data objects
}
\description{
The number of data objects includes all data objects and data stores of a BPMN diagram
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
n_data_objects(file_path)
}
