% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{n_duplicate_tasks}
\alias{n_duplicate_tasks}
\title{Duplicate tasks}
\usage{
n_duplicate_tasks(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the number of duplicate tasks
}
\description{
Duplicate tasks are tasks which share the same name with other tasks
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
n_duplicate_tasks(file_path)
}
