% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{n_message_flows}
\alias{n_message_flows}
\title{Number of message flows}
\usage{
n_message_flows(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the number of message flows
}
\description{
Number of message flows. Message flows are used for communication between processes and link message events
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
n_message_flows(file_path)
}
