% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{n_pools}
\alias{n_pools}
\title{Number of pools}
\usage{
n_pools(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the number of pools
}
\description{
Number of pools in the process models. A pool represents an organisation or an entity
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
n_pools(file_path)
}
