% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{n_swimlanes}
\alias{n_swimlanes}
\title{Number of swimlanes}
\usage{
n_swimlanes(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the number of swimlanes
}
\description{
Number of swimlanes in the pools. A swimlane represents a person, role or team
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
n_swimlanes(file_path)
}
