% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{size_process_model}
\alias{size_process_model}
\title{Size}
\usage{
size_process_model(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the size
}
\description{
The size of a model is the number of tasks, gateways and events
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
size_process_model(file_path)
}
