% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Behavioral_profile.R
\name{some_traces_without_activity}
\alias{some_traces_without_activity}
\title{activity sometimes not in traces}
\usage{
some_traces_without_activity(repetition_and_path_log, xml_internal_doc,
  activity)
}
\arguments{
\item{repetition_and_path_log}{repetition and path log list object created by the function create_repetition_and_path_log}

\item{xml_internal_doc}{document object created using the create_internal_document function}

\item{activity}{the activity name}
}
\value{
a boolean value indicating whether it is true on whether or not an activity is sometimes not part of a trace
}
\description{
{ This functions returns true or false on whether or not an activity is sometimes not part of a trace
This can be useful for measuring the understandability using behavioral profiles.
}
}
\examples{
\dontrun{some_traces_without_activity(log, doc, "A")}
}
