% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{task_names}
\alias{task_names}
\title{Task names}
\usage{
task_names(xml_internal_doc, filter_non_connector_activities = FALSE,
  signavio = FALSE)
}
\arguments{
\item{xml_internal_doc}{document object created using the create_internal_document function}

\item{filter_non_connector_activities}{attribute indicating whether non connector activities should be filtered. The default value is FALSE.}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an object containing a table with the IDs and tasknames
}
\description{
A function which returns the task names together with the task ids
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
task_names(create_internal_doc(file_path))
}
