% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{token_split}
\alias{token_split}
\title{Token Split}
\usage{
token_split(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the token_split
}
\description{
{Token split is defined as the sum of the outgoing flows of parallel, inclusive and complex gateways minus one,
because otherwise the token_split value is always one, while it should be zero if there are}
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
token_split(file_path)
}
