% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnderstandBPMN.R
\docType{package}
\name{understandBPMN}
\alias{understandBPMN}
\alias{understandBPMN-package}
\title{understandBPMN - understandability metrics for BPMN models}
\description{
This package provides the implementation of several comprehensibility and complexity metrics for BPMN models
}
