% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{connector_heterogeneity}
\alias{connector_heterogeneity}
\title{Connector heterogeneity}
\usage{
connector_heterogeneity(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the connector heterogeneity
}
\description{
Connector heterogeneity is defined as the sum of minus - p times the log of p of all gateways. p is defined as the number of a particular type of gateway divided by all gateways.
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
connector_heterogeneity(file_path)
}
