% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{connector_mismatch}
\alias{connector_mismatch}
\title{Connector mismatch}
\usage{
connector_mismatch(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the connector mismatch
}
\description{
Connector mismatch is the absolute value of the difference between split gateways and join gateways for each type of gateway, ie parallel, exclusive, inclusive, complex and event based gateways
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
connector_mismatch(file_path)
}
