% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cross_connectivity}
\alias{cross_connectivity}
\title{Cross Connectivity}
\usage{
cross_connectivity(file_path, signavio = FALSE,
  path_log_already_created = FALSE, generate_new_path_log = FALSE,
  time_to_generate_path_log = 1500)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}

\item{path_log_already_created}{boolean which indicates whether the path log has already been created before or not. When you are not sure, it is best to use the standard which is false}

\item{generate_new_path_log}{used when it is not possible to save the path log such as with the Rapid miner or in unit tests and examples}

\item{time_to_generate_path_log}{time which is the maximum time to generate a new path log in seconds. The standard setting is 1500 seconds.}
}
\value{
an integer indicating the cross connectivity of a model
}
\description{
The cross-connectivity metric that measures the strength of the links between process model elements.
The definition of this new metric builds on the hypothesis that process models are easier understood and contain less errors if they have a high cross-connectivity.
The metric is calculated based on the creation of a data frame containing the values of all connections
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
cross_connectivity(file_path, generate_new_path_log = TRUE)
}
