% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{n_empty_sequence_flows}
\alias{n_empty_sequence_flows}
\title{Empty sequence flows}
\usage{
n_empty_sequence_flows(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the number of empty sequence flows
}
\description{
Empty sequence flow is defined as a flow which connects a split parallel gateway with a join parallel gateway without any tasks in between
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
n_empty_sequence_flows(file_path)
}
