% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sequentiality}
\alias{sequentiality}
\title{Sequentiality}
\usage{
sequentiality(file_path, signavio = FALSE)
}
\arguments{
\item{file_path}{document object created using the create_internal_document function}

\item{signavio}{boolean which indicates whether the file stems from signavio}
}
\value{
an integer indicating the sequentiality
}
\description{
Sequentiality is defined as the number of sequence flows connecting two tasks divided by the total number of sequence flows
}
\examples{
\dontshow{file_path <- system.file("extdata", "doc.txt", package="understandBPMN")}
sequentiality(file_path)
}
